refact: move editExpenseDraft to common expense file
