+++ server/graphql/common/expenses.ts
@@ -65,6 +65,7 @@ import {
   RecipientAccount as BankAccountPayoutMethodData,
   Transfer as WiseTransfer,
 } from '../../types/transferwise';
+import { createUser } from '../common/user';
 import {
   BadRequest,
   FeatureNotAllowedForUser,
@@ -1298,6 +1299,61 @@ export const isAccountHolderNameAndLegalNameMatch = (accountHolderName: string, 
     })
   );
 };
+
+export async function editExpenseDraft(
+  req: express.Request,
+  expenseData: ExpenseData,
+  { args }: { args?: Record<string, any> } = {},
+) {
+  // It is a submit on behalf being completed
+  let existingExpense = await models.Expense.findByPk(expenseData.id, {
+    include: [{ model: models.Collective, as: 'collective' }],
+  });
+  if (!existingExpense) {
+    throw new NotFound('Expense not found.');
+  }
+  if (existingExpense.status !== statuses.DRAFT) {
+    throw new Unauthorized('Expense can not be edited.');
+  }
+  if (existingExpense.data.draftKey !== args.draftKey) {
+    throw new Unauthorized('You need to submit the right draft key to edit this expense');
+  }
+
+  const payeeExists = args.expense.payee?.id || args.expense.payee?.legacyId;
+  const options = { overrideRemoteUser: undefined, skipPermissionCheck: true };
+  if (!payeeExists) {
+    const { organization: organizationData, ...payee } = args.expense.payee;
+    const { user, organization } = await createUser(
+      {
+        ...pick(payee, ['email', 'name', 'legalName', 'newsletterOptIn']),
+        location: expenseData.payeeLocation,
+      },
+      {
+        organizationData,
+        throwIfExists: true,
+        sendSignInLink: true,
+        redirect: `/${existingExpense.collective.slug}/expenses/${expenseData.id}`,
+        creationRequest: {
+          ip: req.ip,
+          userAgent: req.header?.['user-agent'],
+        },
+      },
+    );
+    expenseData.fromCollective = organization || user.collective;
+    options.overrideRemoteUser = user;
+    options.skipPermissionCheck = true;
+  }
+
+  existingExpense = await editExpense(req, expenseData, options);
+
+  await existingExpense.update({
+    status: options.overrideRemoteUser?.id ? statuses.UNVERIFIED : undefined,
+    lastEditedById: options.overrideRemoteUser?.id || req.remoteUser?.id,
+    UserId: options.overrideRemoteUser?.id || req.remoteUser?.id,
+  });
+
+  return existingExpense;
+}
 
 export async function editExpense(req: express.Request, expenseData: ExpenseData, options = {}): Promise<Expense> {
   const remoteUser = options?.['overrideRemoteUser'] || req.remoteUser;
